feat: make CustomUserManager more customizable
Now it is possible to define any field to be used in login, instead of username. Before, you were forced to use email.

Besides that, for compatibility issues the models.py module has been removed. Now, the documentation under manager.py explains how to get the most out of the custom manager when defining a user model.